build(crud-be-csharp-aspnetcore): enable AnalysisLevel=latest-All + Sonar error gate
Add <AnalysisLevel>latest-All</AnalysisLevel> to Directory.Build.props and
escalate all analyzer diagnostics to error in .editorconfig (SonarAnalyzer.CSharp
+ NetAnalyzers build-breaking alongside the existing TreatWarningsAsErrors).

Clean the ~700-finding backlog with real fixes (CA1305 InvariantCulture,
ASP0025 AddAuthorizationBuilder, CA1716 parameter renames, IDE0305 collection
expressions, file-scoped namespaces, unused-using removal). A new
tests/.editorconfig relaxes library/test-idiom rules (CA2007, CA1707, CA2234, …)
while keeping correctness rules at error. Sonar enforcement verified via an
unused-field S1144 break, then removed.

Part of the lint-safety-parity effort (D3).

+++ apps/crud-be-csharp-aspnetcore/tests/DemoBeCsas.Tests/Integration/Steps/AttachmentSteps.cs
@@ -31,7 +31,7 @@ public class AttachmentSteps(
             contentType,
             data
         );
-        ((int)response.StatusCode).Should().Be(
+        response.StatusCode.Should().Be(
             201,
             $"Failed to upload attachment: {response.Body}"
         );
